requirements.txt
numpy==2.1.2
Pillow==10.4.0
scipy==1.14.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.4
